Putting check_skylake function in the x86 directory
